ResumeParser/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain-google-genai
langchain-community
langchain
pydantic[email]
pypdf
python-dotenv